2. Development/Backend/Pont_Mbale/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.8.1
attrs==23.2.0
click==8.1.7
colorama==0.4.6
Django==3.2
djangorestframework==3.12.0
drf-spectacular==0.27.2
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
importlib-resources==6.4.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pkgutil-resolve-name==1.3.10
pydantic==2.3.0
pydantic-core==2.6.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sniffio==1.3.0
sqlparse==0.5.0
starlette==0.27.0
typing-extensions==4.7.1
uritemplate==4.1.1
uvicorn==0.23.2
zipp==3.19.2
